+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## 6.1.0 - 2024-09-22 🫙
 
--   Add "General > showOutputView" setting to toggle automatically showing the output view when running a script ([#492](https://github.com/mark-wiemer-org/ahkpp/issues/492))
+-   Add "General > `showOutputView`" setting to toggle automatically showing the output view when running a script ([#492](https://github.com/mark-wiemer-org/ahkpp/issues/492))
 -   Fix quick help when selecting quoted strings ([#376](https://github.com/mark-wiemer-org/ahkpp/issues/376))
 -   Fix description for V2 > General > `actionWhenV1Detected` config key